Move the logic to create address and derived public key to ChainAccountAddressRepository
