+++ app/src/main/java/com/vultisig/wallet/ui/models/ChainSelectionViewModel.kt
@@ -3,23 +3,16 @@ package com.vultisig.wallet.ui.models
 import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.vultisig.wallet.chains.MayaChainHelper
-import com.vultisig.wallet.chains.PublicKeyHelper
 import com.vultisig.wallet.data.on_board.db.VaultDB
 import com.vultisig.wallet.data.repositories.ChainAccountAddressRepository
-import com.vultisig.wallet.models.Chain
 import com.vultisig.wallet.models.Coin
 import com.vultisig.wallet.models.Coins
-import com.vultisig.wallet.models.TssKeysignType
 import com.vultisig.wallet.models.Vault
-import com.vultisig.wallet.tss.TssKeyType
 import com.vultisig.wallet.ui.navigation.Screen.VaultDetail.AddChainAccount
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
-import wallet.core.jni.PublicKey
-import wallet.core.jni.PublicKeyType
 import javax.inject.Inject
 
 internal data class ChainSelectionUiModel(
@@ -53,56 +46,16 @@ internal class ChainSelectionViewModel @Inject constructor(
             commitVault(checkNotNull(vaultDb.select(vaultId)) {
                 "No vault with $vaultId"
             }.let { vault ->
-                // When user add a coin to the vault , we need to derive the public key and address
-                // and save the address and public key with the coin , thus don't need to derive in the future
-                when (coin.chain.TssKeysignType) {
-                    TssKeyType.ECDSA -> {
-                        val derivedPublicKey = PublicKeyHelper.getDerivedPublicKey(
-                            vault.pubKeyECDSA,
-                            vault.hexChainCode,
-                            coin.coinType.derivationPath()
-                        )
-                        if (coin.chain == Chain.mayaChain) {
-                            vault.copy(
-                                coins = vault.coins + coin.copy(
-                                    address = MayaChainHelper(
-                                        vault.pubKeyECDSA,
-                                        vault.hexChainCode
-                                    ).getCoin()?.address ?: "",
-                                    hexPublicKey = derivedPublicKey
-                                )
-                            )
-                        } else {
-                            val address = chainAccountAddressRepository.getAddress(
-                                coin.coinType,
-                                PublicKey(
-                                    derivedPublicKey.hexToByteArray(),
-                                    PublicKeyType.SECP256K1
-                                )
-                            )
-                            vault.copy(
-                                coins = vault.coins + coin.copy(
-                                    address = address,
-                                    hexPublicKey = derivedPublicKey
-                                )
-                            )
-                        }
-                    }
-
-                    TssKeyType.EDDSA -> {
-                        val address = chainAccountAddressRepository.getAddress(
-                            coin.coinType,
-                            PublicKey(vault.pubKeyEDDSA.hexToByteArray(), PublicKeyType.ED25519)
-                        )
-                        vault.copy(
-                            coins = vault.coins + coin.copy(
-                                address = address,
-                                hexPublicKey = vault.pubKeyEDDSA
-                            )
-                        )
-                    }
-                }
-
+                val (address, derivedPublicKey) = chainAccountAddressRepository.getAddress(
+                    coin,
+                    vault
+                )
+                vault.copy(
+                    coins = vault.coins + coin.copy(
+                        address = address,
+                        hexPublicKey = derivedPublicKey
+                    )
+                )
             })
         }
     }
